+++ deployment/modules/hetzner/templates/docker-compose.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+echo "Running: docker compose $@"
+
+CURRENT_UID=$(id -u):$(id -g) docker compose $@