+++ deployment/modules/hetzner/variables.tf
@@ -1,3 +1,12 @@
+variable "veebot_tg_image_tag" {
+  type = string
+}
+
+variable "tg_bot_token" {
+  type      = string
+  sensitive = true
+}
+
 variable "prometheus_remote_write_url" {
   type = string
 }
@@ -11,7 +20,7 @@ variable "prometheus_password" {
   sensitive = true
 }
 
-variable "loki_remote_write_url" {
+variable "loki_url" {
   type = string
 }