+++ deployment/prod/main.tf
@@ -1,13 +1,16 @@
 module "hetzner" {
   source = "../modules/hetzner"
 
+  tg_bot_token        = var.tg_bot_token
+  veebot_tg_image_tag = var.veebot_tg_image_tag
+
   prometheus_remote_write_url = grafana_cloud_stack.this.prometheus_remote_write_endpoint
   prometheus_username         = grafana_cloud_stack.this.prometheus_user_id
   prometheus_password         = var.grafana_cloud_api_key
 
-  loki_remote_write_url = "${grafana_cloud_stack.this.logs_url}/loki/api/v1/push"
-  loki_username         = grafana_cloud_stack.this.logs_user_id
-  loki_password         = var.grafana_cloud_api_key
+  loki_url      = grafana_cloud_stack.this.logs_url
+  loki_username = grafana_cloud_stack.this.logs_user_id
+  loki_password = var.grafana_cloud_api_key
 }
 
 module "grafana_cloud_stack" {