+++ veebot-telegram/src/config.rs
@@ -1,5 +1,7 @@
 use serde::{de::DeserializeOwned, Deserialize};
+use std::collections::HashMap;
 use teloxide::types::UserId;
+use tracing_subscriber::prelude::*;
 
 pub struct Config {
     pub(crate) tg: TgConfig,
@@ -23,18 +25,61 @@ pub(crate) struct DbConfig {
 }
 
 fn default_database_pool_size() -> u32 {
-    // Free Postgres instances hosted on Heroku have 20 connections limit.
-    // However, we also reserve 1 connection for ad-hoc db administration purposes
+    // Postgres instance has 100 connections limit.
+    // However, we also reserve 2 connections for ad-hoc db administration purposes
     // via pg_admin, for example.
-    19
+    98
 }
 
 impl Config {
     pub fn load_or_panic() -> Config {
         Self {
-            tg: from_env_or_panic("TELEGRAM_"),
+            tg: from_env_or_panic("TG_"),
             db: from_env_or_panic("DATABASE_"),
         }
+    }
+}
+
+#[derive(Deserialize)]
+pub struct LoggingConfig {
+    loki_url: url::Url,
+    loki_username: String,
+    loki_password: String,
+    #[serde(with = "serde_with::json::nested")]
+    veebot_log_labels: HashMap<String, String>,
+}
+
+impl LoggingConfig {
+    pub fn load_or_panic() -> LoggingConfig {
+        from_env_or_panic("")
+    }
+
+    pub fn init_logging(self) {
+        let env_filter = tracing_subscriber::EnvFilter::from_env("VEEBOT_LOG");
+
+        let fmt = tracing_subscriber::fmt::layer()
+            .with_target(true)
+            .with_ansi(std::env::var("COLORS").as_deref() != Ok("0"))
+            .pretty();
+
+        let mut loki_url = self.loki_url.clone();
+        loki_url.set_username(&self.loki_username).unwrap();
+        loki_url.set_password(Some(&self.loki_password)).unwrap();
+
+        let mut labels = self.veebot_log_labels;
+
+        labels.insert("source".to_owned(), "veebot".to_owned());
+
+        let (loki, task) =
+            tracing_loki::layer(loki_url, labels, HashMap::new()).unwrap();
+
+        tokio::spawn(task);
+
+        tracing_subscriber::registry()
+            .with(fmt)
+            .with(loki)
+            .with(env_filter)
+            .init();
     }
 }