+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Move the bot to Hetzner Cloud
 - Scrape the server with grafana-agent sending telemetry to Grafana Cloud
+    - The Linux server is configured with node_exporter only
 - Add a script to SSH to Hetzner server
 - Add a script to deploy the bot on Hetzner server and upgrade the docker image
 - Define full infrastructure with terraform that include the deployment of the following:
@@ -11,3 +12,6 @@
     - Grafana Cloud Stack
     - Hetzner server
 - Add .editorconfig
+- Add `cargo xtask` utility for development in docker-compose environment
+- Add `tracing_loki` layer to the applicating to send logs to Grafana Cloud
+- Add a NAS volume for Hetzner server to use as persistent storage for future Postgres instance